backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.0
spidev==3.6
